requirements.txt
Flask==3.0.0
google-generativeai
gunicorn==21.2.0